+++ js/fetch.js
@@ -1,26 +1,33 @@
-import { createMiniature } from './create-miniature.js';
-
-const body = document.querySelector('body');
-
-const errorContainer = document.querySelector('#error').content;
-const errorTemplate = errorContainer.querySelector('section');
-const errorMessage = errorTemplate.cloneNode(true);
-const showErrorMessage = () => {
-  body.appendChild(errorMessage);
+const getData = (onSuccess, onError) => {
+  fetch('https://24.javascript.pages.academy/kekstagram/data')
+    .then((response) => response.json())
+    .then((photos) => {
+      onSuccess(photos);
+    });
+    .catch (() => {
+  onError();
+};
 };
 
-const successContainer = document.querySelector('#success').content;
-const successTemplate = successContainer.querySelector('section');
-const successMessage = successTemplate.cloneNode(true);
-const showSuccessMessage = () => {
-  body.appendChild(successMessage);
+const sendData = (onSuccess, onError, data) => {
+  fetch('https://24.javascript.pages.academy/kekstagram/data',
+    {
+      method: 'POST',
+      data,
+    },
+  )
+    .then((response) => {
+      if (response.ok) {
+        onSuccess();
+      } else {
+        onError;
+      }
+    })
+}
+  .catch (() => {
+  onError();
+};
 };
 
-fetch('https://24.javascript.pages.academy/kekstagram/data')
-  .then((response) => response.json())
-  .then((photos) => {
-    createMiniature(photos);
-  })
-  .catch(() => {
-    showErrorMessage();
-  });
+
+export { getData, sendData };
