+++ js/main.js
@@ -1,11 +1,16 @@
 import { arrayPhotos } from './make-array-photos.js';
 // import { createMiniature } from './create-miniature.js';
 import { showFullPicture } from './show-full-photo.js';
+import { showErrorMessage, showSuccessMessage } from './fetch-messages.js';
 import './form.js';
 import './edit-scale.js';
 import './slider.js';
-import './fetch.js';
+import { createMiniature } from './create-miniature.js';
+import { getData } from './fetch.js';
+import './submit-button.js';
 
 // createMiniature(arrayPhotos);
 
 showFullPicture(arrayPhotos[2]);
+
+getData(createMiniature, showErrorMessage);
